refactor: 인게임과 메뉴 패키지 분리 #62
