+++ src/main/java/org/teamseven/tetris/ui/MainFrame.java
@@ -1,6 +1,7 @@
 package org.teamseven.tetris.ui;
 
 import org.teamseven.tetris.Pipeline;
+import org.teamseven.tetris.ui.menu.MainMenuPane;
 
 import static org.teamseven.tetris.Const.*;
 
@@ -12,7 +13,7 @@ public class MainFrame extends BaseFrame {
 
     @Override
     public void setComp() {
-        getContentPane().add(new GameMenuPane());
+        getContentPane().add(new MainMenuPane());
     }
 
     @Override
